requirements.txt
pillow>=10.0.0
numpy>=1.21.0
matplotlib>=3.5.0
# Optional: kagglehub>=0.1.0 (for --kaggle flag)